pyproject.toml
[project]
name = "slm-serving-challenge1-weaction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "structlog>=25.5.0",
    "uvicorn>=0.42.0",
]
